requirements.txt
requests~=2.28.1
toml~=0.10.2
Pillow~=9.3.0
python-dotenv~=0.21.0
py-cord